refactoring: JUnit4 & JUnit5 test classes moved into separate packages
Renamed test classes back to 'CashbackHackServiceTest'. Now each of them belongs to its own package

Provided some basic logging facility `LOGGER`
